+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 * `unnest` maintains order
 * `column` index access
+* array column scalars
 
 ## [2.0.2](https://pypi.org/project/graphique/2.0.2/) - 2026-01-10
 ### Changed